requirements.txt
requests
beautifulsoup4
html2text
google-genai
rich
